Always show search form on Special:Search
At least for wikis that have AdvancedSearch installed it does not
seem to make sense to hide the interface. It should be usefull in
any case. Also there's a huge gap that should be avoided I guess.

I'm not to sure for cases where AdvancedSearch is not installed.
There this patch will lead to a situation where there are two
search fields visible. The one at the top and the one in the main
content. - Although the latter has the advantage to show the
current search term.

IMO this is overall still an improvement and further streamlining
could be done in a follow up ( like hiding the header search field
generally on the Special:Search page ).

Bug: T308845
Change-Id: I477825e9d738e9934f6c6845a1ff8962f12987a9