Omit the last modified bar in the HTML rather than hiding it via CSS
The mediawiki.diff.styles can be loaded on pages other than the diff
page itself, so it is better to move this logic into SkinMinerva and
avoid outputting it for diff pages full stop.

Note this module is loaded on Special:MobileDiff and desktop
Minerva history page (without redirect) so we need to cater
for both cases.

Bug: T350515
Change-Id: If2d908b7cbe257184e0c899f78a04d4d5935ea8c