Show diff switcher on narrow screens
The visual diff is sometimes cluttered at narrow widths, but as
we show the switcher on tablet and desktop, and it stores a preference,
users may see the visual diff on phones anyway.

The layout is sometimes cluttered but not completely useless, and
users can always switch back.

As a follow-up we should improve the layout of the visual diff sidebar
on narrow screens.

Bug: T351946
Change-Id: I5533126195331b46040e708b7b35627e74a3001d