Fix mobile diff author note on dark mode
Updates the styles for the author note shown at the bottom of the screen
on the diff page on narrow displays to use Codex design tokens so that
it will use dark colors when dark mode is enabled.

The color for the edit count is changed from #72777d (color-placeholder
/ color-disabled) to a darker color #54595d to use a more appropriate
token (color-subtle).

Bug: T370003
Change-Id: I3340e8794036a1c7cd8aa4cfebd2ecfb50d2cbcf