Disable unused icon variants
By default this module was inheriting variants and since it doesnt declare a
property selectorWithVariant it was using the default selectorWithVariant
rule which impacts OOUI icons.

Explicitly disable the variants for this module as they are unused. This
reduces the size of the modules.

Follow up to: I87b060cf98194d81679da2610944f58e8d941389

Bug: T374145
Bug: T376226
Change-Id: Icc008449ee498a946066c4ce5d46adcac194bd3b