pageactions: Show page actions on Main page diff
Currently all page actions are not shown for main page view-action
(which can be read-view or diff-view) and so they're not generated
in the first place.

To be able to show them on `diff-view` and not show them on `read-view`,
we have to generate them unconditonally first on view action, and then
either show or hide them with CSS as appropriate.

Also add test talk button on main page.

Bug: T363916
Change-Id: Ica660fa57ac55cae534bc03cc68b0381e07fd47d