Drop message box styling modules from page load
The skins.minerva.messageBox.styles module is loaded on pages with
old revisions, but this message box is already being styled by another
module added by core.

Skins are not responsible for styling Codex. This is the responsibility
of the code that adds the message box to the page.

In addition to this we were loading another set of styles for message boxes
on skins.minerva.codex.styles

Bug: T375287
Change-Id: I0cf5dc6baf2e447b4944f04365addd614d59cfbe