Remove transparent PNG icon handling
The existing code prevents a black square from appearing on late loading icons.
The problem with it is that if loaded late, it can override actual icon definitions
and make the icon invisible.

Since this typically applies in the case of icons loaded via JavaScript
this shouldn't be a problem on mobile web

Bug: T266361
Change-Id: Icee701761e9db69a6bffb7af4f608ec26c30f2be