Bottom dock: Add support
Copy across from Vector 2022 so that it can be used consistently across
our two default skins. The menu by default is icon-only (no text labels)

The new menu can be extended by PHP hook or mw.util.addPortletLink

The addPortletLink hook is updated to check for the presence of the
class minerva-icon-only-menu to decide whether the menu is icon
only.

Bug: T384733
Change-Id: I76dfc06d3529d00f393cd618ea4a815d9531803f