Fix logic for when to show the notification circle
This functionality was seemingly broken in I55c18cf72.

The comments in the code declare that the notification circle
should show when there are unseen "notifications" (where a
"notification" is either an "alert" on a "notice), and this
matches user expectations. However the logic as implemented
means we only see the circle when there are unseen "alerts".

Bug: T344029
Change-Id: I91cd5e7703760323c4a6786e8c7a7a00b21e8de2
(cherry picked from commit acb5cb08bfc43e7f43717dba35399fe5233c90d0)