Re-apply: "Fix logic for when to show the notification circle"
This functionality was seemingly broken in I681a2540979 / I55c18cf72.

The comments in the code declare that the notification circle
should show when there are unseen "notifications" (where a
"notification" is either an "alert" on a "notice"), and this
matches user expectations. However the logic as implemented
means we only see the circle when there are unseen "alerts".

Bug: T344029
Change-Id: I22743b4cff04670ed3279a307e89adea84601a9d