Styles: Workaround for Codex 2.0 release
(A similar workaround is also required in Vector).

The upcoming Codex 2.0 release changes the name of a mixin
file used in Minerva. This patch is needed to get around a
"chicken-and-egg" situation where the patch in MW core
(which introduces the new Codex files, including one with
a changed name) fails tests because Minerva expects the old
file to still be present.

This patch temporarily comments-out the offending code in Minerva
in a way that should not have any user-facing impact (because
all the light-mode values are provided as fallbacks).

A follow-up patch in Minerva after the Codex 2.0 update in Core
will undo these changes and update the Less import to the new
filename.

Bug: T391012
Change-Id: Ib4ab3729f2329f435a2d85bc5f8cda617f165ee5