Remove overflow:hidden from #mw-mf-page-center
This rule was added to:
* hide a horizontal scrollbar caused
by popup menus (personal tools, overflow menu) being partially
offscreen. Those tools used to be full on screen, which
removes the need to clip the whole page.
* address a markup issue in the checkbox hack which didn't work with
instrumentation
*  ensure the skin is always responsive, even when content it not

Clipping the whole page caused a regression in DiscussionTools,
and is moved to the article content child element.

The markup of the dropdown is updated so that it doesn't cause
the page to overflow.

Bug: T395937
Change-Id: I0433a0c653af9bdc5f1154395c19172b29ae7d73