Table width should not exceed content width
On narrow screens, width: 100% results in tables taking up the
entire width of the page. This is undesirable for smaller tables
and results in visual anomalies.

This patch sets table width to fit its content. Note that if
max-width is smaller than width, the computed width value is
always overridden by max-width, so there is no risk of having
an oversized table.

Bug: T402066
Change-Id: If38a07f9dadea7fce86ae413b56e6b1d04e331ca