Only promote watchstar from actions if bookmark is not present
The watchstar is typically bundled in the "actions" menu which is
part of the page tools. Minerva promotes it to actions just like
the watchstar. Going forward we are exploring replacing this with
a bookmark button.

For now, detect the presence of the bookmark icon and only promote
it if it's not there.

Bug: T402940
Change-Id: I0f88bb17466f1677c310da264b3cb46468a6f84a