Default main menu should consider MinervaPersonalMenu configuration
With tests document the behaviour of $wgMinervaPersonalMenu and
$wgMinervaAdvancedMainMenu and how they are tied. This change
also makes it possible to enable $wgMinervaPersonalMenu without
enabling $wgMinervaAdvancedMainMenu meaning in future we can
consider removing some of this logic and we can enable the
personal tools for logged in users without the advanced hamburger
menu.

Changes:
* To support unit testing, I move from SpecialPage to the associated
SpecialPageFactory service function which is equivalent.
* A RuntimeException is thrown in unsupported combinations of config
rather than relying on notes in README.md

Bug: T404139
Change-Id: I2c3e8991ee9259e42fc7f4722f25f954e9dc5a4c