Removal of AMC outreach drawer, Minerva side
Removes use of MobileFrontend's AMC outreach campaign setup
and callout drawer, as it's not been working in some time and
we don't need to keep dead code around.

Should be merged before or together with the MobileFrontend-side
changes. Safe to deploy this one without the MobileFrontend patch,
but may be unsafe to deploy the MobileFrontend patch without
this one.

Bug: T402470
Change-Id: I3a543e26bcb3ad155912c94700c151ba34aea66d